refactor: 重構 API 服務，將型別定義移至新檔案，並精簡不必要的型別，提升代碼可讀性與維護性。
